backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pandas==2.2.3
scikit-learn==1.5.2
joblib==1.4.2
numpy==1.26.4
gunicorn==23.0.0
